refactor: move chunk files to chunks dir
