+++ src/chunks/tiny_chunk.c
@@ -1,4 +1,4 @@
-#include "../inc/main.h" 
+#include "../../inc/main.h" 
 
 
 void print_tiny_chunk(t_tiny_chunk* tiny) {
@@ -10,7 +10,7 @@ t_tiny_chunk* create_top_tiny_chunk(t_fpage* page) {
     t_tiny_chunk* tiny = (t_tiny_chunk*)FASTPAGE_SHIFT(page);
     tiny->size = page->chunk_size;
     tiny->next = NULL;
-    char *data = (void*)MEMORY_SHIFT(tiny, sizeof(size_t));
+    char* data = (char*)MEMORY_SHIFT(tiny, sizeof(size_t));
 
     log_info("creating top tiny chunk");
     printf("tiny chunk pointer: %p\n", tiny);
